두 개의 정수를 입력받아 두 정수 사이(두 정수를 포함)에 3의 배수이거나 5의 배수인 수들의 합과 평균을 출력하는 프로그램을 작성하시오.

+++ src/algorithm/PrimeString.java
@@ -15,7 +15,7 @@ public class PrimeString {
             primeTable[x] = 0;
         } // initial PrimeTable
         for(int i = 0; i < inputLength; i++) {
-            int ascii = ((int) input.charAt(i)) - 65;
+            int ascii = (input.charAt(i)) - 65;
             primeTable[ascii] += 1;
         }
